+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Ethereum Gas Tracker Changelog
 
 ## [Initial Version] - 2023-04-24
+ - Add Show Current Gas Price in Ethereum Mainnet